requirements.txt
python-dotenv
openai
flask
pypdf2